Fix table height and reuse setColumnWidth
