+++ app/javascript/app/components/ndcs/lts-explore-table/lts-explore-table.js
@@ -5,6 +5,7 @@ import { withRouter } from 'react-router';
 import PropTypes from 'prop-types';
 import qs from 'query-string';
 import { getLocationParamUpdated } from 'utils/navigation';
+import { setColumnWidth } from 'utils/table';
 
 import { actions as fetchActions } from 'pages/lts-explore';
 
@@ -48,21 +49,9 @@ class LTSExploreTableContainer extends PureComponent {
   }
 
   setColumnWidth = column => {
-    const narrowColumns = [0, 1];
-    const tableWidth = 1170;
-    const numColumns = this.props.columns.length;
-    const numNarrowColumns = narrowColumns.length;
-    const colPadding = 10;
-    const narrowColumnWidth = 180;
-    const columnWidth =
-      (tableWidth -
-        (numColumns + 2) * colPadding -
-        numNarrowColumns * narrowColumnWidth) /
-      (numColumns - numNarrowColumns);
-    const index = this.props.columns.indexOf(column);
-    return index !== -1 && narrowColumns.indexOf(index) > -1
-      ? narrowColumnWidth
-      : columnWidth;
+    const TABLE_WIDTH = 1170;
+    const MIN_COLUMN_WIDTH = 180;
+    setColumnWidth(column, this.props.columns, TABLE_WIDTH, MIN_COLUMN_WIDTH);
   };
 
   updateUrlParam(param, clear) {
