+++ app/javascript/app/components/ndcs/ndcs-explore-table/ndcs-explore-table-selectors.js
@@ -86,9 +86,23 @@ const headerChanges = {
   'Share of GHG Emissions': 'Share of global GHG emissions'
 };
 
+export const getDefaultColumns = createSelector(
+  [getIndicatorsParsed],
+  indicators => {
+    if (!indicators || isEmpty(indicators)) return [];
+    const columnIds = ['country', 'submission', 'submission_date', 'ndce_ghg'];
+
+    const columns = columnIds.map(id => {
+      const match = indicators.find(indicator => indicator.value === id);
+      return match ? match.label : id;
+    });
+    return columns.map(c => headerChanges[c] || c);
+  }
+);
+
 export const tableRemoveIsoFromData = createSelector(
-  [tableGetFilteredData],
-  data => {
+  [tableGetFilteredData, getDefaultColumns],
+  (data, columnHeaders) => {
     if (!data || isEmpty(data)) return null;
     return data.map(d => {
       const updatedTableDataItem = { ...d };
@@ -118,27 +132,16 @@ export const tableRemoveIsoFromData = createSelector(
         `/ndcs/country/${updatedTableDataItem.iso}` +
         "'>"}${updatedTableDataItem.country}</a>`;
       delete updatedTableDataItem.iso;
-      const changedHeadersD = {};
+
+      const filteredAndChangedHeadersD = {};
       Object.keys(updatedTableDataItem).forEach(k => {
         const header = headerChanges[k] || k;
-        changedHeadersD[header] = updatedTableDataItem[k];
+        if (columnHeaders.includes(header)) {
+          filteredAndChangedHeadersD[header] = updatedTableDataItem[k];
+        }
       });
-      return changedHeadersD;
+      return filteredAndChangedHeadersD;
     });
-  }
-);
-
-export const getDefaultColumns = createSelector(
-  [getIndicatorsParsed],
-  indicators => {
-    if (!indicators || isEmpty(indicators)) return [];
-    const columnIds = ['country', 'submission', 'submission_date', 'ndce_ghg'];
-
-    const columns = columnIds.map(id => {
-      const match = indicators.find(indicator => indicator.value === id);
-      return match ? match.label : id;
-    });
-    return columns.map(c => headerChanges[c] || c);
   }
 );
